VideoController.deleteVideo: Unit Test 2 Complete

+++ spooktube-api/src/controllers/Video.controller.js
@@ -73,7 +73,12 @@ export default class VideoController {
     
     async deleteVideo(req, res) {
         try {
-            await this.#videoService.checkOwnership(req.body.videoId, req.body.userId);
+            const ownedVideo = await this.#videoService.checkOwnership(req.body.videoId, req.body.userId);
+            
+            if (ownedVideo === null) {
+                return res.status(401).json({ message: "User does not own this video" });
+            }
+            
             await this.#videoService.deleteVideo(req.body.videoId);
             await this.#contentManagerService.deleteVideo(req.body.videoId);
